Moved io.kuzzle.sdk.core to io.kuzzle.sdk.test
