+++ src/main/java/io/kuzzle/sdk/test/Kuzzle.java
@@ -1,4 +1,4 @@
-package io.kuzzle.sdk.core;
+package io.kuzzle.sdk.test;
 
 import android.util.Log;
